requirements.txt
attrs==25.3.0
boto3==1.37.31
botocore==1.37.31
certifi==2025.1.31
charset-normalizer==3.4.1
Deprecated==1.2.18
eodhp-utils @ git+https://github.com/UKEODHP/eodhp-utils.git@e10b425e8546fe53f1f7eae9b9c09e5fdc212a28
Faker==37.1.0
idna==3.10
importlib_metadata==8.6.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
multidict==6.4.2
opentelemetry-api==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-logging==0.52b1
opentelemetry-processor-baggage==0.52b1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
packaging==24.2
pulsar-client==3.6.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
s3transfer==0.11.4
six==1.17.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
wrapt==1.17.2
zipp==3.21.0
